Extract a module for the slugging operations.
