+++ lib/kilt/base.rb
@@ -11,3 +11,4 @@ require "kilt/db/active_record"
 require "kilt/utils"
 require "kilt/form"
 require "kilt/formatting"
+require "kilt/slugger"
